added value expression instruction creation service fixed float atom object internal value type

+++ src/cz/mg/application/entities/buildin/atoms/float32/MgFloat32BinaryOperator.java
@@ -29,11 +29,11 @@ public abstract class MgFloat32BinaryOperator extends MgBinaryOperator implement
     public void run(MgTask task) {
         MgFloat32Object leftObject = (MgFloat32Object) task.getObject(getLeft());
         MgFloat32Object rightObject = (MgFloat32Object) task.getObject(getRight());
-        int left = leftObject.getValue();
-        int right = rightObject.getValue();
-        int result = compute(left, right);
+        float left = leftObject.getValue();
+        float right = rightObject.getValue();
+        float result = compute(left, right);
         task.setObject(getResult(), new MgFloat32Object(result));
     }
 
-    protected abstract int compute(int left, int right);
+    protected abstract float compute(float left, float right);
 }
